Fix edge case for msgid_plural without comments
A small quirk of using #maybe from Parslet is that the value might be
added to the previous element instead of having its own. To avoid this,
we use #repeat with a range of 0 to 1 (most likely, an outcome from the
fact that a nil value for #maybe is nil while for #repeat is an empty
array).